wrap code in public function
